Add releases and Docker support.
